Add first draft of a user report
Gets date from the following sources

- GetMailbox (Exchange Online PowerShell Module)
- Get-ADUser for in-premise AD
- GetUser for Office 365 only accounts

Fixed issue with dot sourcing of /ps/ files as it was dot sourcing
files that were not valid .ps1 scripts

Also added type names to custom classes to enable formatting in tables.
As part of that changed the name of Office class to OBNOffice.